Correct type annotation for the index

+++ src/main.rs
@@ -70,7 +70,7 @@ fn search_worker(paths_mutex_queue: GlobalPathQueue) {
 }
 
 fn paths_worker(paths_queue: GlobalPathQueue, dir_paths_queue: GlobalPathQueue, 
-    still_finding_paths: Arc<Vec<AtomicBool>>, idx: u32) {
+    still_finding_paths: Arc<Vec<AtomicBool>>, idx: usize) {
     loop {
         if still_finding_paths.iter().all(|b| !b.load(Ordering::Relaxed)) {
             break;
